+++ changelog.md
@@ -27,3 +27,6 @@
 
 ## v3.0.0
 * fallback to UTC, instead of PST if no `Intl` is present
+* support passing-in offsets as ISO_8601 date-strings
+* add epoch-seconds warning msg
+* allow getting/setting new timezones